Conflicts:
	CREDITS
	ChangeLog
	Makefile
	Makefile.conf.template
	Makefile.defs
	cfg.y
	db/schema/entities.xml
	doc/dbschema/xsl/pi_framework_mod.xsl
	etc/opensips.cfg
	evi/event_interface.c
	main.c
	modules/b2b_logic/README
	modules/cachedb_couchbase/cachedb_couchbase_dbase.c
	modules/cachedb_memcached/cachedb_memcached.c
	modules/db_http/http_dbase.c
	modules/dialog/dlg_handlers.c
	modules/dispatcher/dispatch.c
	modules/drouting/drouting.c
	modules/event_rabbitmq/event_rabbitmq.c
	modules/httpd/README
	modules/imc/README
	modules/imc/imc.c
	modules/imc/imc_cmd.c
	modules/imc/imc_mng.c
	modules/mi_http/http_fnc.c
	modules/pi_http/README
	modules/pi_http/http_fnc.c
	modules/presence/publish.c
	modules/pua/hash.c
	modules/pua/pua.c
	modules/ratelimit/ratelimit_helper.c
	modules/registrar/reg_mod.c
	modules/registrar/save.c
	modules/rtpproxy/README
	modules/rtpproxy/rtpproxy.c
	modules/siptrace/siptrace.c
	modules/uac_registrant/registrant.c
	packaging/debian-lenny/changelog
	packaging/debian/changelog
	packaging/fedora/opensips.spec
	packaging/freebsd/Makefile
	packaging/gentoo/opensips-1.11.5.ebuild
	packaging/gentoo/opensips-1.9.0.ebuild
	packaging/gentoo/opensips-1.9.1.ebuild
	packaging/netbsd/Makefile
	packaging/openbsd/Makefile
	packaging/rpm/opensips.spec.CentOS
	packaging/rpm/opensips.spec.SuSE
	packaging/solaris/base-pkginfo
	packaging/solaris/berkeley-pkginfo
	packaging/solaris/carrierroute-pkginfo
	packaging/solaris/identity-pkginfo
	packaging/solaris/ldap-pkginfo
	packaging/solaris/mmgeoip-pkginfo
	packaging/solaris/mysql-pkginfo
	packaging/solaris/perl-pkginfo
	packaging/solaris/pgsql-pkginfo
	packaging/solaris/regex-pkginfo
	packaging/solaris/snmp-pkginfo
	packaging/solaris/tls-pkginfo
	packaging/solaris/xmlrpc-pkginfo
	scripts/opensipsdbctl.base
	scripts/pi_http/acc-mod
	scripts/pi_http/cachedb_sql-mod
	scripts/pi_http/carrierroute-mod
	scripts/pi_http/dialog-mod
	scripts/pi_http/dispatcher-mod
	scripts/pi_http/drouting-mod
	scripts/pi_http/group-mod
	scripts/pi_http/imc-mod
	scripts/pi_http/pi_framework.xml
	scripts/pi_http/presence-mod
	scripts/pi_http/registrar-mod
	scripts/pi_http/rls-mod
	scripts/pi_http/usrloc-mod
	sctp_server.c
	tcp_main.c

+++ modules/statistics/statistics.c
@@ -17,8 +17,8 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License 
- * along with this program; if not, write to the Free Software 
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
  * History:
@@ -95,7 +95,7 @@ static pv_export_t mod_items[] = {
 
 struct module_exports exports= {
 	"statistics",		/* module's name */
-	MODULE_VERSION,	
+	MODULE_VERSION,
 	DEFAULT_DLFLAGS,	/* dlopen flags */
 	cmds,				/* exported functions */
 	mod_params,			/* param exports */
@@ -145,7 +145,7 @@ static int fixup_stat(void** param, int param_no)
 		/* reference to the statistic name */
 		sp = (struct stat_param *)pkg_malloc(sizeof(struct stat_param));
 		if (sp==NULL) {
-			LM_ERR("no more pkg mem (%ld)\n",sizeof(struct stat_param));
+			LM_ERR("no more pkg mem (%d)\n", (int)sizeof(struct stat_param));
 			return E_OUT_OF_MEM;
 		}
 		memset( sp, 0 , sizeof(struct stat_param) );
@@ -425,7 +425,7 @@ int pv_set_stat(struct sip_msg* msg, pv_param_t *param, int op,
 
 	if (val != 0)
 		LM_WARN("non-zero value - setting value to 0\n");
-	
+
 	reset_stat( stat );
 
 	return 0;
